🔧 Turn component to basic component for redesign in feature.
